Add platform (#45440)
* Add platform

This will be helpful when parts of the code need to take different
actions based on the platform Istio is running on.

Having this set as a global value allows it to be used across different
charts and components.

This is a cheap way of achieving this ability - It involves no runtime
check - aka calls to the API server to distinguish between different
distributions of Kubernetes.

* Use empty value to mean vanilla Kubernetes

* Use Default instead of Empty